Arcade game with sound and responsive design.
Characters are animated. Sound files added to index.html and played in
app.js file. Responsive design created by using hammer.js for touch
control and adding window resize event listener that resizes the canvas
to fit nicely on whatever is the size of the browser screen.
Manifest.json file was added to allow the web app to be installed to an
android home screen.